+++ src/Config/Config.php
@@ -63,14 +63,14 @@ class Config
         $options = [
             'metadata.broker.list' => $this->broker,
             'auto.offset.reset' => config('kafka.offset_reset', 'latest'),
-            'enable.auto.commit' => config('kafka.auto_commit', 'true'),
+            'enable.auto.commit' => config('kafka.auto_commit', true) === true ? 'true' : 'false',
             'compression.codec' => config('kafka.compression', 'snappy'),
             'group.id' => $this->groupId,
             'bootstrap.servers' => $this->broker,
         ];
 
-        if ($this->autoCommit) {
-            $options['enable.auto.commit'] = 'true';
+        if (isset($this->autoCommit)) {
+            $options['enable.auto.commit'] = $this->autoCommit === true ? 'true' : 'false';
         }
 
         return array_merge($options, $this->customOptions);